Add a ReplaceType dep mapper and move them into their own file
